+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Bug fixes:
 * Fix keeping `Float` and big `Integer` elements of an `Array` converted to native storage by `RARRAY_PTR()` alive (@eregon).
 * Fix `Marshal.load` for an object extended with a module containing a nested user-marshaled object (#3943, @andrykonchin).
 * Fix `Regexp` comments in extended mode to always end at a newline, even when the newline is preceded by a backslash (#4435, @eregon).
+* Fix issue where environment updates made in native code (e.g., a C extension calling `setenv(3)`) would not be reflected by `ENV` (@nirvdrum).
 
 Compatibility: